requirements.txt
flask==2.0.1
flask-socketio==5.1.1
flask-login==0.5.0
flask-sqlalchemy==2.5.1
google-generativeai==0.3.1
python-dotenv==0.19.1
requests==2.31.0
gunicorn==20.1.0
bcrypt==4.0.1
Pillow==10.0.0
python-dateutil==2.8.2
APScheduler==3.10.1 